test: split board persistence scenarios and fixtures
